+++ apps/faces-example2/src/main/java/org/apache/struts/webapp/example2/LoggedOff.java
@@ -26,8 +26,6 @@ package org.apache.struts.webapp.example2;
 import java.io.IOException;
 
 import javax.faces.FacesException;
-import javax.faces.bean.ManagedBean;
-import javax.faces.bean.RequestScoped;
 import javax.faces.context.FacesContext;
 
 import org.slf4j.Logger;
@@ -38,8 +36,6 @@ import org.slf4j.LoggerFactory;
  * <p>Backing bean for the <code>loggedoff.jsp</code> page.</p>
  */
 
-@ManagedBean
-@RequestScoped
 public class LoggedOff {